# Performance optimization: Instantiate Chips only once