--- a/src/gadgets/wnn.rs
+++ b/src/gadgets/wnn.rs
@@ -24,7 +24,6 @@ pub trait WnnInstructions<F: PrimeFieldBits> {
     fn predict(
         &self,
         layouter: impl Layouter<F>,
-        bloom_filter_arrays: Array3<bool>,
         inputs: Vec<F>,
     ) -> Result<Vec<AssignedCell<F, F>>, Error>;
 }
@@ -43,22 +42,40 @@ pub struct WnnChipConfig<F: PrimeFieldBits> {
 }
 
 /// Implements a BTHOWeN- style weightless neural network.
-/// 
+///
 /// This happens in three steps:
 /// 1. The [`HashChip`] is used to range-check and hash the inputs.
 /// 2. The [`BloomFilterChip`] is used to look up the bloom filter responses
 ///    (for each input and each class).
 /// 3. The [`ResponseAccumulatorChip`] is used to accumulate the responses.
 struct WnnChip<F: PrimeFieldBits> {
-    config: WnnChipConfig<F>,
-    _marker: PhantomData<F>,
+    hash_chip: HashChip<F>,
+    bloom_filter_chip: BloomFilterChip<F>,
+    response_accumulator_chip: ResponseAccumulatorChip<F>,
+
+    n_classes: usize,
+    n_inputs: usize,
 }
 
 impl<F: PrimeFieldBits> WnnChip<F> {
-    fn construct(config: WnnChipConfig<F>) -> Self {
+    fn construct(config: WnnChipConfig<F>, bloom_filter_arrays: Array3<bool>) -> Self {
+        let n_classes = bloom_filter_arrays.shape()[0];
+        let n_inputs = bloom_filter_arrays.shape()[1];
+
+        let hash_chip = HashChip::construct(config.hash_chip_config.clone());
+        let bloom_filter_chip = BloomFilterChip::construct(
+            config.bloom_filter_chip_config.clone(),
+            bloom_filter_arrays,
+        );
+        let response_accumulator_chip =
+            ResponseAccumulatorChip::construct(config.response_accumulator_chip_config.clone());
+
         WnnChip {
-            config,
-            _marker: PhantomData,
+            hash_chip,
+            bloom_filter_chip,
+            response_accumulator_chip,
+            n_classes,
+            n_inputs,
         }
     }
 
@@ -96,49 +113,33 @@ impl<F: PrimeFieldBits> WnnChip<F> {
             response_accumulator_chip_config,
         }
     }
+
+    pub fn load(&mut self, layouter: &mut impl Layouter<F>) -> Result<(), Error> {
+        self.bloom_filter_chip.load(layouter)
+    }
 }
 
 impl<F: PrimeFieldBits> WnnInstructions<F> for WnnChip<F> {
     fn predict(
         &self,
         mut layouter: impl Layouter<F>,
-        bloom_filter_arrays: Array3<bool>,
         inputs: Vec<F>,
     ) -> Result<Vec<AssignedCell<F, F>>, Error> {
-        assert_eq!(bloom_filter_arrays.shape()[1], inputs.len());
-
-        // Flatten array: from shape (C, N, B) to (C * N, B)
-        let shape = bloom_filter_arrays.shape();
-        let bloom_filter_arrays_flat = bloom_filter_arrays
-            .clone()
-            .into_shape((shape[0] * shape[1], shape[2]))
-            .unwrap();
-
-        let hash_chip = HashChip::construct(self.config.hash_chip_config.clone());
-        let mut bloom_filter_chip = BloomFilterChip::<F>::construct(
-            self.config.bloom_filter_chip_config.clone(),
-            bloom_filter_arrays_flat,
-        );
-        let response_accumulator_chip = ResponseAccumulatorChip::<F>::construct(
-            self.config.response_accumulator_chip_config.clone(),
-        );
+        assert_eq!(self.n_inputs, inputs.len());
 
         let hashes = inputs
             .iter()
-            .map(|input| hash_chip.hash(layouter.namespace(|| "hash"), *input))
+            .map(|input| self.hash_chip.hash(layouter.namespace(|| "hash"), *input))
             .collect::<Result<Vec<_>, _>>()?;
 
-        let n_classes = bloom_filter_arrays.shape()[0];
-        bloom_filter_chip.load(&mut layouter)?;
-
         let mut responses = vec![];
-        for c in 0..n_classes {
+        for c in 0..self.n_classes {
             responses.push(Vec::new());
-            for (i, hash) in hashes.clone().iter().enumerate() {
+            for (i, hash) in hashes.clone().into_iter().enumerate() {
                 let array_index = c * hashes.len() + i;
-                responses[c].push(bloom_filter_chip.bloom_lookup(
+                responses[c].push(self.bloom_filter_chip.bloom_lookup(
                     &mut layouter,
-                    hash.clone(),
+                    hash,
                     F::from(array_index as u64),
                 )?);
             }
@@ -147,7 +148,8 @@ impl<F: PrimeFieldBits> WnnInstructions<F> for WnnChip<F> {
         responses
             .iter()
             .map(|class_responses| {
-                response_accumulator_chip.accumulate_responses(&mut layouter, class_responses)
+                self.response_accumulator_chip
+                    .accumulate_responses(&mut layouter, class_responses)
             })
             .collect::<Result<Vec<_>, _>>()
     }
@@ -270,11 +272,12 @@ impl<F: PrimeFieldBits> Circuit<F> for WnnCircuit<F> {
         config: Self::Config,
         mut layouter: impl halo2_proofs::circuit::Layouter<F>,
     ) -> Result<(), halo2_proofs::plonk::Error> {
-        let wnn_chip = WnnChip::construct(config.wnn_chip_config);
+        let mut wnn_chip =
+            WnnChip::construct(config.wnn_chip_config, self.bloom_filter_arrays.clone());
+        wnn_chip.load(&mut layouter)?;
 
         let result = wnn_chip.predict(
             layouter.namespace(|| "wnn"),
-            self.bloom_filter_arrays.clone(),
             self.inputs.iter().map(|v| F::from(*v)).collect(),
         )?;