Profiling the current code a little bit, I noticed that the performance optimization in #12 introduced a lot of duplicate computation (in particular, bloom filter bits were packed into field elements over and over again). As a result, circuit synthesis used up around half the proving time!

I fixed this by switching to a pattern where, if a chip uses another chip, it should instantiate it already in the constructor. That way, if a chip is used many times, any one-time computation can happen in the constructor.

The benchmark results compared to `main` branch:

```
benches/proof_generation
                        time:   [1.0382 s 1.0674 s 1.1003 s]
                        change: [-51.766% -49.847% -47.765%] (p = 0.00 < 0.05)
                        Performance has improved.
benches/key_generation  time:   [622.13 ms 655.07 ms 688.86 ms]
                        change: [-76.549% -75.192% -73.880%] (p = 0.00 < 0.05)
                        Performance has improved.
benches/verification    time:   [13.310 ms 14.287 ms 15.066 ms]
                        change: [-4.4483% +4.0829% +13.279%] (p = 0.40 > 0.05)
                        No change in performance detected.
```

(Note that I started using a different computer today, so these numbers are not comparable to numbers in previous PRs. I re-ran the benchmark on `main` though and the reported percentages are correct.)